pass along the context of the console error to the faro error
