+++ packages/web-sdk/src/instrumentations/console/instrumentation.ts
@@ -1,11 +1,11 @@
 import {
   allLogLevels,
   BaseInstrumentation,
-  defaultLogArgsSerializer,
-  LogArgsSerializer,
   LogLevel,
   VERSION,
 } from '@grafana/faro-core';
+
+import { getDetailsFromErrorArgs } from '../../utils';
 
 import type { ConsoleInstrumentationOptions } from './types';
 
@@ -14,7 +14,6 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
   readonly version = VERSION;
 
   static defaultDisabledLevels: LogLevel[] = [LogLevel.DEBUG, LogLevel.TRACE, LogLevel.LOG];
-  private errorSerializer: LogArgsSerializer = defaultLogArgsSerializer;
 
   constructor(private options: ConsoleInstrumentationOptions = {}) {
     super();
@@ -23,7 +22,6 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
   initialize() {
     this.logDebug('Initializing\n', this.options);
     this.options = { ...this.options, ...this.config.consoleInstrumentation };
-    this.errorSerializer = this.options.errorSerializer ?? defaultLogArgsSerializer;
 
     allLogLevels
       .filter(
@@ -34,7 +32,9 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
         console[level] = (...args) => {
           try {
             if (level === LogLevel.ERROR && !this.options?.consoleErrorAsLog) {
-              this.api.pushError(new Error('console.error: ' + this.errorSerializer(args)));
+              const { value, type, stackFrames } = getDetailsFromErrorArgs(args);
+
+              this.api.pushError(new Error('console.error: ' + value), { type, stackFrames });
             } else {
               this.api.pushLog(args, { level });
             }
